Ruby 2.4 and UI improvements (#122)
* refactor (UI): method actions on layout and other changes

* feat (ruby): updates to ruby 2.4.0 and all gems

* refactor (files): deletes unused files

* fix (UI): more UI updates on layout links

* fix (UI): more improvements

* refactor (ruby): 2.4.0 deploy updates

* fix (nginx): updates nginx config to ruby 2.4

* fix (stats): boxing issues fixed for stats page

* fix (gtags): adds balance to gtags index

+++ db/migrate/20170203162158_rename_fees_in_event.rb
@@ -1,8 +1,12 @@
 class RenameFeesInEvent < ActiveRecord::Migration[5.0]
   def change
-    remove_column :events, :gtag_deposit_fee, :integer
-    rename_column :events, :card_return_fee, :initial_topup_fee
-    rename_column :events, :gtag_deposit, :gtag_deposit_fee
+    remove_column :events, :gtag_deposit_fee, :integer if column_exists?(:events, :gtag_deposit_fee, :integer)
+    remove_column :events, :card_return_fee, :integer if column_exists?(:events, :card_return_fee, :integer)
+    remove_column :events, :gtag_deposit, :integer if column_exists?(:events, :gtag_deposit, :integer)
+
+    add_column :events, :initial_topup_fee, :integer, default: 0 unless column_exists?(:events, :initial_topup_fee, :integer)
+    add_column :events, :gtag_deposit_fee, :integer, default: 0 unless column_exists?(:events, :gtag_deposit_fee, :integer)
+
     change_column :events, :initial_topup_fee, :integer, default: 0
 
     Event.find_by_sql("UPDATE events SET initial_topup_fee = 0")
